Implemented Json2Save for all other SaveParts  - Fully working for example save  - Encryption of recreated save still missing

+++ WolfSave/SavePart7.h
@@ -57,7 +57,7 @@ protected:
 
 	void json2Save(JsonReader &jr, FileWriter &fw) const
 	{
-		DWORD var1 = jr.Read<BYTE>();
+		BYTE var1 = jr.Read<BYTE>();
 		fw.Write(var1);
 
 		if (var1 == 1)
